+++ core/initializer/parallelism/test_executor_parallelizer.go
@@ -1,9 +1,15 @@
 package parallelism
 
 import (
+	"fmt"
 	"github.com/docker/distribution/uuid"
 	"github.com/docker/docker/client"
+	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
+	"io"
+	"io/ioutil"
+	"os"
+	"strings"
 	"sync"
 )
 
@@ -68,12 +74,11 @@ Args:
 	allTestParams: A mapping of test_name -> parameters for running the test
 
 Returns:
-	A mapping of test_name -> test output info
+	True if all tests passed, false otherwise
  */
-func (executor TestExecutorParallelizer) RunInParallel(interceptor *ErroneousSystemLogCaptureWriter, allTestParams map[string]ParallelTestParams) map[string]ParallelTestOutput {
+func (executor TestExecutorParallelizer) RunInParallel(allTestParams map[string]ParallelTestParams) bool {
 	// These need to be buffered else sending to the channel will be blocking
 	testParamsChan := make(chan ParallelTestParams, len(allTestParams))
-	testOutputChan := make(chan ParallelTestOutput, len(allTestParams))
 
 	logrus.Info("Loading test params into work queue...")
 	for _, testParams := range allTestParams {
@@ -82,34 +87,32 @@ func (executor TestExecutorParallelizer) RunInParallel(interceptor *ErroneousSys
 	close(testParamsChan) // We close the channel so that when all params are consumed, the worker threads won't block on waiting for more params
 	logrus.Info("All test params loaded into work queue")
 
+	outputManager := newParallelTestExecutionOutputManager()
+
 	logrus.Infof("Launching %v tests with parallelism %v...", len(allTestParams), executor.parallelism)
-	executor.disableSystemLogAndRunTestThreads(interceptor, testParamsChan, testOutputChan)
+	executor.disableSystemLogAndRunTestThreads(outputManager, testParamsChan)
 	logrus.Info("All tests exited")
 
-	// Collect all results
-	result := make(map[string]ParallelTestOutput)
-	for i := 0; i < len(allTestParams); i++ {
-		output := <-testOutputChan
-		result[output.TestName] = output
-	}
-	return result
+	outputManager.printSummary()
+	return outputManager.getAllTestsPassed()
 }
 
-func (executor TestExecutorParallelizer) disableSystemLogAndRunTestThreads(interceptor *ErroneousSystemLogCaptureWriter, testParamsChan chan ParallelTestParams, testOutputChan chan ParallelTestOutput) {
+func (executor TestExecutorParallelizer) disableSystemLogAndRunTestThreads(
+		outputManager *ParallelTestExecutionOutputManager,
+		testParamsChan chan ParallelTestParams) {
 	/*
     Because each test needs to have its logs written to an independent file to avoid getting logs all mixed up, we need to make
     sure that all code below this point uses the per-test logger rather than the systemwide logger. However, it's very difficult for
     a coder to remember to use 'log.Info' when they're used to doing 'logrus.Info'. To enforce this, we capture any systemwide logger usages
 	during this function so we can show them later.
 	*/
-	currentSystemOut := logrus.StandardLogger().Out
-	logrus.SetOutput(interceptor)
-	defer logrus.SetOutput(currentSystemOut)
+	outputManager.startInterceptingStdLogger()
+	defer outputManager.stopInterceptingStdLogger()
 
 	var waitGroup sync.WaitGroup
 	for i := uint(0); i < executor.parallelism; i++ {
 		waitGroup.Add(1)
-		go executor.runTestWorkerGoroutine(&waitGroup, testParamsChan, testOutputChan)
+		go executor.runTestWorkerGoroutine(outputManager, &waitGroup, testParamsChan)
 	}
 	waitGroup.Wait()
 }
@@ -119,18 +122,31 @@ A function, designed to be run inside a worker thread, that will pull test param
 push the result to the test results channel
  */
 func (executor TestExecutorParallelizer) runTestWorkerGoroutine(
+			outputManager *ParallelTestExecutionOutputManager,
 			waitGroup *sync.WaitGroup,
-			testParamsChan chan ParallelTestParams,
-			testOutputChan chan ParallelTestOutput) {
+			testParamsChan chan ParallelTestParams) {
 	// IMPORTANT: make sure that we mark a thread as done!
 	defer waitGroup.Done()
 
 	for testParams := range testParamsChan {
-		// Create a separate logger just for this test that writes to the given file
+		testName := testParams.TestName
+
+		tempFilename := fmt.Sprintf("%v-%v", executor.executionId, testName)
+		writingTempFp, err := ioutil.TempFile("", tempFilename)
+		if err != nil {
+			emptyOutputReader := &strings.Reader{}
+			executionErr := stacktrace.Propagate(err, "An error occurred creating temporary file to contain logs of test %v", testName)
+			outputManager.logTestOutput(testName, executionErr, false, emptyOutputReader)
+			continue
+		}
+		defer os.Remove(writingTempFp.Name())
+
+		// Create a separate logger just for this test that writes to a tempfile
 		log := logrus.New()
 		log.SetLevel(logrus.GetLevel())
-		log.SetOutput(testParams.LogFp)
+		log.SetOutput(writingTempFp)
 		log.SetFormatter(logrus.StandardLogger().Formatter)
+
 		testExecutor := newTestExecutor(
 			log,
 			executor.executionId,
@@ -139,17 +155,22 @@ func (executor TestExecutorParallelizer) runTestWorkerGoroutine(
 			executor.testControllerImageName,
 			executor.testControllerLogLevel,
 			executor.customTestControllerEnvVars,
-			testParams.TestName,
+			testName,
 			testParams.Test)
 
 		passed, executionErr := testExecutor.runTest()
+		writingTempFp.Close() // Close to flush out anything remaining in the buffer
 
-		testOutput := ParallelTestOutput{
-			TestName:     testParams.TestName,
-			ExecutionErr: executionErr,
-			TestPassed:   passed,
-			LogFp:        testParams.LogFp,
+		// Create a new FP to read the logfile from the start
+		var testOutputReader io.Reader
+		readingTempFp, err := os.Open(writingTempFp.Name())
+		if err != nil {
+			errorMsg := fmt.Sprintf("An error occurred opening the test's logfile for reading; logs for this test are unavailable:\n%s", err)
+			testOutputReader = strings.NewReader(errorMsg)
+		} else {
+			defer readingTempFp.Close()
+			testOutputReader = readingTempFp
 		}
-		testOutputChan <- testOutput
+		outputManager.logTestOutput(testName, executionErr, passed, testOutputReader)
 	}
 }