+++ core/CHANGELOG.md
@@ -1,4 +1,5 @@
 # TBD
+* Print test output as the tests finish, rather than waiting for all tests to finish to do so
 
 # 0.8.0
 * Simplify service config definition to a single method