+++ core/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Change ConfigurationID to be a string
 * Print test output as the tests finish, rather than waiting for all tests to finish to do so
 * Gracefully clean up tests when SIGINT, SIGQUIT, or SIGTERM are received
+* Tiny bugfix in printing test output as tests finish
 
 # 0.8.0
 * Simplify service config definition to a single method