+++ core/CHANGELOG.md
@@ -1,4 +1,4 @@
-# TBD
+# 0.9.0
 * Change ConfigurationID to be a string
 * Print test output as the tests finish, rather than waiting for all tests to finish to do so
 * Gracefully clean up tests when SIGINT, SIGQUIT, or SIGTERM are received